Removed dummy SPSET construct. Setting of SP local variables is now part of the existing SET syntax.
Note: This has the result that a somewhat extended syntax (from SQL-99)
is allowed. We allow a list of settings ("SET a=1, b=2, ...;"), where the
different variables can be of different types (SP local, system or user (@)).
This also means that certain optional modifiers, such as GLOBAL, are allowed
when setting an SP local variable, but then has no meaning and is ignored.


sql/lex.h:
  Removed dummy SPSET symbol.
sql/sql_yacc.yy:
  Removed dummy SPSET construct.
  Setting of SP local variables is now part of the existing SET syntax.