Due to a compiler bug, slave.cc:tables_ok() sometimes wrongly returns 1 if the return type is int or int_fast8_t.  The test case that showed this problem is rpl000001 and the tested version was MySQL 5.0.2.  The compiler with the problem is GCC 3.0.4 runing on "Linux bitch 2.4.18 #2 Thu Apr 11 14:37:17 EDT 2002 sparc64 unknown".
By changing the return type to bool the problem disappear.  (Another
way to make the problem disappear is to simply print the returned
value with printf("%d",?).  The printed returned value is always 0 in
the test cases I have run.)  This is only a partial solution to the
problem, since someone could later change the return type of the
function back to int or some other type that does not work.


sql/slave.cc:
  Changed type
sql/slave.h:
  Changed type