Impl4 of WL2278 - Dynamic port allocation of cluster nodes
When a node restarts it starts over again with fetching the configuration
-  It is not sure that it can use the "old dynamically allocated port number" again.
-  It should however try to reuse the old one, if not possible it should
allocate a new one. One has to be able to distinguish between portnumbers
specified originally in the config, and ones that has been dynamically added
(the latter may be changed if "busy", but the first cannot be changed).

We use negative portnumbers for ports that are ok to change.