WL2278 - Dynamic port allocation for cluster nodes
Re-introduce ability to have
[tcp default]PortNumber= 4242

in config.ini to NOT use dynamic ports.

Dynamic port numbers are the default though.

We still use the mgm port for transporter connections to a mgm server though.
This is (slightly) different than before, but since you have to be able to connect
to the mgm server anyway, this does not require extra firewall rules or anything.


ndb/src/mgmsrv/ConfigInfo.cpp:
  Restore (optional) old behaviour of having a base port number and incrementing.