1.2048 05/10/20 10:31:16 ingo@mysql.com +1 -0 Bug#12166 - Unable to index very large tables Moved clearing of errors behind the second repair attempt, but will clear only if no error happened.
No test case. The error can be repeated with little free
space on tmpdir only. I do not know of a way to create this
in a portable way with our test suite.

I did however attach a shell script to the bug report which
can easily be adapted to the situation on the test machine.


myisam/mi_check.c:
  Bug#12166 - Unable to index very large tables
  Changed comments.
  Removed unnecessary use of param->keys_in_use.
sql/ha_myisam.cc:
  Bug#12166 - Unable to index very large tables
  Moved clearing of errors behind the second repair attempt,
  but will clear only if no error happened.