Bug#18129 - Fast (online) add index leaves temporary table frm in case of errors
ALTER TABLE temporarily creates a new table with a .frm
file and optionally other files. For fast ALTER TABLE
only the .frm file is created. If the operation succeeds,
The temporary files are renamed to their final target.

In case of an error, the temporary file was forgotten to
remove.

Manually tested. The test requires to look at files,
which I think cannot be done portably with the test suite.
The test file is attached to the bug report.