Re-apply missing changesets in 5.1
BUG#19544 mysqldump does not backup TS and LG information correctly.
  
Problem was the type of the column was varchar(8) not an integer type. So mysqld
was outputting a string representation of the number in scientific notation,
which wasn't exact. Now we do it as a bigint and output the number in bytes.

Fix assertion when loading plugins due to handler changes